requirements.txt
asgiref
cffi
confusable-homoglyphs
cryptography
Django
django-allauth
django-registration
dotenv
fido2
gunicorn
packaging
pycparser
python-dotenv
sqlparse
psycopg2-binary
psycopg
qrcode
bs4
requests
django-webpush
icalendar
sentry-sdk[django]
yt-dlp
sendgrid
django-recaptcha
google-auth